+++ Source/Physics/GenericBasicSolver.cpp
@@ -325,7 +325,7 @@ uint GenericBasicSolver::FindRigidBodyIndex(Collider* collider)
   uint notFoundVal = uint(-1);
   RigidBody* body = collider->GetActiveBody();
   uint index = mLookupMap.FindValue(body, notFoundVal);
-  uint objectCount = static_cast<uint>(mObjects.Size());
+  uint objectCount = mObjects.Size();
   if (index == notFoundVal)
   {
     ConstraintObjectData objData;
