+++ Source/Physics/PhysicsMeshBoundData.hpp
@@ -166,7 +166,7 @@ public:
       DoNotifyException("Invalid Range Operation", "Cannot access an item in an empty range.");
       return FrontResult();
     }
-    return mArray->Get(static_cast<uint>(mIndex));
+    return mArray->Get(mIndex);
   }
 
   void PopFront()
