+++ Source/Physics/PhysicsSolverConfig.cpp
@@ -447,7 +447,7 @@ void PhysicsSolverConfig::RebuildConstraintBlockValues()
   mJointBlocks.Resize(JointEnums::JointCount);
   for (size_t i = 0; i < mJointBlocks.Size(); ++i)
   {
-    mJointBlocks[i].mJointId = static_cast<uint>(i);
+    mJointBlocks[i].mJointId = i;
     mJointBlocks[i].ResetDefaultValues();
   }
   mContactBlock.mJointId = JointEnums::JointCount;
@@ -499,7 +499,7 @@ void PhysicsSolverConfig::CopyTo(PhysicsSolverConfig* destination)
 
 uint PhysicsSolverConfig::GetSize() const
 {
-  return static_cast<uint>(mBlocks.Size());
+  return mBlocks.Size();
 }
 
 HandleOf<ConstraintConfigBlock> PhysicsSolverConfig::GetBlockAt(uint index)
